requirements.txt
streamlit>=1.28.0
docling
ollama>=0.1.5
chromadb>=0.4.15
sentence-transformers>=2.2.2
pandas>=2.1.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
tqdm>=4.66.1
